requirements.txt
pynetworktables
websockets
